refactor(blocks): extract insert inline latex to command (#8594)
This PR extract the implementation of insert inline equation from slash menu to a new rich-text command named `insertInlineLatex`
