+++ packages/affine/components/src/rich-text/effects.ts
@@ -3,6 +3,7 @@ import type { deleteTextCommand } from './format/delete-text.js';
 import type { formatBlockCommand } from './format/format-block.js';
 import type { formatNativeCommand } from './format/format-native.js';
 import type { formatTextCommand } from './format/format-text.js';
+import type { insertInlineLatex } from './format/insert-inline-latex.js';
 import type {
   getTextStyle,
   isTextStyleActive,
@@ -64,6 +65,7 @@ declare global {
       toggleLink: typeof toggleLink;
       getTextStyle: typeof getTextStyle;
       isTextStyleActive: typeof isTextStyleActive;
+      insertInlineLatex: typeof insertInlineLatex;
     }
   }
 }
